In this update, I have added the ability to attempt automatic rebracketing of your system. This is not flawless at the moment, but it should offer the ability to more quickly view results that would otherwise have been cut off at box boundaries.
Why it's not flawless: Right now this handles rebracketing by determining the number of beads contained within slices along each axis
of length 0.01, with the lowest-density slice (or the first slice that contains no beads, if one exists) being chosen for the box boundary in
that direction. The far wall of the box will be pressed forward by that amount, and distances past 0.5 will be reduced by 1. For example, if the
slice in the x-direction from 0.47 to 0.48 contained no beads, the x-axis boundaries would be 0.47 to 1.47. If, in a different system, no slice were empty of beads, but the lowest-density slice along the x-axis occurred at 0.63 to 0.64, the box x-axis boundaries would be -0.37 to 0.63.

The problem with this is that the real metric for finding the best box boundary isn't the *bead* density, but rather the *bond* density - i.e., how many bonds are severed by a plane if that distance is chosen as the box boundary. These two metrics are not always identical, so it's not uncommon to see some small amount of "orphaned" beads and/or bonds on one end of the box or the other.

I may add a more robust (but markedly slower) option to attempt box rebracketing by bonds instead of by beads, but this is not a priority for the moment.